Created action mapper and moved mapping methods from action repository to it
